+++ src/SchemaTrigger.ts
@@ -2,6 +2,7 @@ import { QlikRepositoryClient } from "qlik-rest-api";
 import { ISchemaEvent, ITaskUpdateTriggerSchema } from "./Task.interface";
 import { IEntityRemove } from "./types/interfaces";
 import { TDaysOfMonth, TDaysOfWeek, TRepeatOptions } from "./types/ranges";
+import { schemaRepeat } from "./util/schemaTrigger";
 
 export interface IClassSchemaTrigger {
   remove(): Promise<IEntityRemove>;
@@ -18,7 +19,7 @@ export class SchemaTrigger implements IClassSchemaTrigger {
     id: string,
     details?: ISchemaEvent
   ) {
-    if (!id) throw new Error(`tags.get: "id" parameter is required`);
+    if (!id) throw new Error(`schemaTrigger.get: "id" parameter is required`);
 
     this.id = id;
     this.repoClient = repoClient;
@@ -50,7 +51,7 @@ export class SchemaTrigger implements IClassSchemaTrigger {
     if (arg.daylightSavingTime)
       this.details.daylightSavingTime = arg.daylightSavingTime ? 1 : 0;
 
-    let schemaRepeatOpt = this.schemaRepeat(
+    let schemaRepeatOpt = schemaRepeat(
       arg.repeat || "Daily",
       arg.repeatEvery || 1,
       arg.daysOfWeek || ["Monday"],
@@ -67,64 +68,5 @@ export class SchemaTrigger implements IClassSchemaTrigger {
     return await this.repoClient
       .Put(`schemaevent/${this.details.id}`, this.details)
       .then((res) => ({ id: this.details.id, status: res.status }));
-  }
-
-  // TODO: repeats in ReloadTaskBase
-  private schemaRepeat(
-    repeat: TRepeatOptions,
-    repeatEvery: number,
-    daysOfWeek: TDaysOfWeek[],
-    daysOfMonth: TDaysOfMonth[]
-  ): { incrementDescr: string; schemaFilterDescr: string } {
-    if (repeat == "Once")
-      return {
-        incrementDescr: "0 0 0 0",
-        schemaFilterDescr: "* * - * * * * *",
-      };
-
-    if (repeat == "Minute")
-      return {
-        incrementDescr: `${repeatEvery} 0 0 0`,
-        schemaFilterDescr: "* * - * * * * *",
-      };
-
-    if (repeat == "Hourly")
-      return {
-        incrementDescr: `0 ${repeatEvery} 0 0`,
-        schemaFilterDescr: "* * - * * * * *",
-      };
-
-    if (repeat == "Daily")
-      return {
-        incrementDescr: `0 0 ${repeatEvery} 0`,
-        schemaFilterDescr: "* * - * * * * *",
-      };
-
-    if (repeat == "Weekly") {
-      let weekDay = this.getWeekDayNumber(daysOfWeek);
-      return {
-        incrementDescr: `0 0 1 0`,
-        schemaFilterDescr: `* * - ${weekDay} ${repeatEvery} * * *`,
-      };
-    }
-
-    if (repeat == "Monthly")
-      return {
-        incrementDescr: `0 0 1 0`,
-        schemaFilterDescr: `* * - * * ${daysOfMonth} * *`,
-      };
-  }
-
-  // TODO: repeats in ReloadTaskBase
-  private getWeekDayNumber(daysOfWeek: TDaysOfWeek[]): number[] {
-    return daysOfWeek.map((d) => {
-      if (d == "Sunday") return 0;
-      if (d == "Monday") return 1;
-      if (d == "Tuesday") return 2;
-      if (d == "Wednesday") return 3;
-      if (d == "Thursday") return 4;
-      if (d == "Friday") return 5;
-      if (d == "Saturday") return 6;
-    });
   }
 }